refactor: separate entities
