+++ src/entities/bill.entity.ts
